Modularize write_ops.rs into submodules for maintainability
Split 524-line write_ops.rs into domain-focused submodules:
- create.rs: issue creation
- comment.rs: commenting
- status.rs: close/reopen operations
- labels.rs: label add/remove
- assign.rs: issue assignment
