+++ src/cli/issues/write_ops/mod.rs
@@ -1,0 +1,14 @@
+//! Issue write operation commands (create, comment, close, reopen, label, assign)
+
+mod assign;
+mod comment;
+mod create;
+mod labels;
+mod status;
+
+// Re-export public commands
+pub use assign::cmd_assign;
+pub use comment::cmd_comment;
+pub use create::cmd_create;
+pub use labels::cmd_label;
+pub use status::{cmd_close, cmd_reopen};
